Rework plugin API, provide methods for late roles
A fairly common want/need for a plugin is to apply a role after
all the other plugins and BEGIN blocks have been applied.  I added
a general mechanism on the main module which runs finalizing code
on the extension process, and convenience methods for plugins to
apply roles or parent classes.

This might be over-engineering, but it will give plugin authors
more options to sort out the kind of problems that come up when
applying roles.